Initial performance optimization
Changed the mzXML reading functions to only decode scans that are needed
later on during the processing, as opposed to decoding all scans.
Additionally, I did some code cleanup in the filenames list, allowing
the user to specify a different format for files that must be aligned as
for files that must be extracted.