Differential extraction and alignment files
The user can now specify a unique string for both extraction and
alignment. This is achieved by checking for the presence of a 'h5' file
prior to the for file in filenames loop, if not present the program will
glob for the files (using the unique strings). Furthermore, a small bug
was fixed that resulted in the uncalibrated sumspectra to be written in
the folder of the LaCyTools file instead of the Batchfolder.