LaCyTools v1.0.1
Implemented a different method for the fitting of the spline to get the
accurate mass, which fixes bug #19 by using the
InterpolatedUnivariateSpline. Furthermore, LaCyTools now comes with
prorgress bars for both alignment and quantitation. Lastly, most of the
output from LaCyTools is now redirected towards the log file and no
longer the shell terminal.