LaCyTools v1.1.0-alpha b20180419a
This is an alpha build for v1.1.0 (b20180419a) that fixes bug #31, by
ensuring that each individual elements isotopic distribution is not cut
pre-maturely (by the if f < 0.01 construction). The fix now checks if
the current fraction is below the threshold and if the fraction is
smaller than the previous fraction.

Furthermore, no new features will be added to this version and build to
prepare it for enhancement #32, which requires that all changeable
parameters are taken out of the source code and put into regular files
that can be read at run-time.