LaCyTools v1.1.0-alpha b20180423a
This build implements a integrity check of the block files that are
available to LaCyTools (located in the blocks folder). Specifically, it
checks if each block contains mass of type float, carbons of type int,
hydrogens of type int, nitrogens of type int, oxygens of type int,
sulfurs of type int and available_for_charge_carrier of type int and a
value of either 0 or 1.

The program will show a message box if any block fails this check, prior
to shutting down.