LaCyTools v1.1.0-alpha b20180508a
This build implements a singular function that creates the headers for
the summary file, which included changes to fix bug #26. Moreso, this
build also fixes the header for the scenario where a user wants to
quantify an analyte at differing retention time windows (but identical
composition and retention time).

This build then unconvered the bugs #36 and #37. The first is caused by
the time window not being evaluated in the actual data section of the
combine results function as the data is correct in the underlying raw
files.