LaCyTools v1.1.0-alpha b20181102b
This build implements a setting (can only be edited in the source code) that tells LaCyTools if it is allowed to create a new analytes.ref file or if it should use the existing analytes.ref file (line 53).

NOTE; This does not check if there actually is an analytes.ref file present in the folder, this is the user's responsability.